Add create new post test
